Adds /parser service with postman tests.
