+++ src/services/parser/parser.js
@@ -1,11 +1,14 @@
 import { ParserService, getOptions } from './parser.class.js'
+import { hooks as schemaHooks } from '@feathersjs/schema'
 import { authorizeHook } from '../../auth/authorize.hook.js'
 import { authenticate } from '@feathersjs/authentication'
+import { parserQueryValidator, parserDataValidator } from './parser.schema.js'
 
 export const parserPath = 'parser'
-export const parserMethods = ['create']
+export const parserMethods = ['create','convert']
 export * from './parser.class.js'
 export * from './parser.schema.js'
+
 
 // A configure function that registers the service and its hooks via `app.configure`
 export const parser = (app) => {
@@ -16,6 +19,7 @@ export const parser = (app) => {
     // You can add additional custom events to be sent to clients here
     events: []
   })
+
   // Initialize hooks
   app.service(parserPath).hooks({
     around: {
@@ -25,10 +29,13 @@ export const parser = (app) => {
       all: [
         // authenticate('googleIAP'),
         // authorizeHook,
+        // schemaHooks.validateQuery(parserQueryValidator)
       ],
       find: [],
       get: [],
-      create: [],
+      create: [
+        // schemaHooks.validateData(parserDataValidator)
+      ],
       patch: [],
       remove: []
     },
